+++ constants/index.js
@@ -1,0 +1,8 @@
+const THIRTY_DAYS = 2592000;
+
+const DISTRIBUTOR = '0xE527972DD80970bb90977879658FB83aD568Bb28';
+
+module.exports = {
+  DISTRIBUTOR,
+  THIRTY_DAYS,
+};
